chore: migrate project to TypeScript
