+++ .gitignore
@@ -8,3 +8,4 @@ example/preferences.json
 .sass-cache
 node_modules
 build
+dist
