+++ src/app/components/main/components/group/components/fields/accelerator/index.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import keycodeToChar from '../../../../../../../utils/keycodeToChar.js';
+import keycodeToChar from '../../../../../../../utils/keycodeToChar';
 
 const AcceleratorField = ({ field, value, onChange }) => {
 
@@ -101,9 +101,9 @@ const AcceleratorField = ({ field, value, onChange }) => {
 
 	return (
 		<div className={`field field-accelerator key-${field.key}`}>
-			<div className="field-label">{ field.label }</div>
-			<input type="text" value={ (pressing && accelerator) || value } aria-label={ field.label } onKeyDown={ handleKeyDown } onKeyUp={ handleKeyUp } readOnly />
-			{ field.help && <span className="help">{ field.help }</span> }
+			<div className='field-label'>{ field.label }</div>
+			<input type='text' value={ (pressing && accelerator) || value } aria-label={ field.label } onKeyDown={ handleKeyDown } onKeyUp={ handleKeyUp } readOnly />
+			{ field.help && <span className='help'>{ field.help }</span> }
 		</div>
 	);
