+++ src/app/components/main/components/group/components/fields/dropdown/index.tsx
@@ -13,12 +13,12 @@ class DropdownField extends React.Component {
 
 		return (
 			<div className={`field field-dropdown key-${this.field.key}`}>
-				<div className="field-label">{ this.label }</div>
+				<div className='field-label'>{ this.label }</div>
 				<select onChange={ this.onChange.bind(this) } value={ this.value } aria-label={ this.label }>
-					<option value="">-- Select One --</option>
+					<option value=''>-- Select One --</option>
 					{ options }
 				</select>
-				{ this.help && <span className="help">{ this.help }</span> }
+				{ this.help && <span className='help'>{ this.help }</span> }
 			</div>
 		);
