+++ src/app/components/main/components/group/components/fields/radio/index.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { newGuid } from '../../../../../../../utils/newGuid.js';
+import { newGuid } from '../../../../../../../utils/newGuid';
 
 class RadioField extends React.Component {
 
@@ -16,10 +16,10 @@ class RadioField extends React.Component {
 			const id = `${fieldID}_${idx}`;
 
 			return (
-				<label htmlFor={ id } className="radio-option" key={ idx }>
+				<label htmlFor={ id } className='radio-option' key={ idx }>
 					{ option.label }
-					<input type="radio" id={ id } onChange={ this.onChange.bind(this) } checked={ option.value === this.value } name={ fieldID } aria-label={ option.label } />
-					<span className="check-circle" />
+					<input type='radio' id={ id } onChange={ this.onChange.bind(this) } checked={ option.value === this.value } name={ fieldID } aria-label={ option.label } />
+					<span className='check-circle' />
 				</label>
 			);
 
@@ -27,9 +27,9 @@ class RadioField extends React.Component {
 
 		return (
 			<div className={`field field-radio key-${this.field.key}`}>
-				<div className="field-label">{ this.label }</div>
+				<div className='field-label'>{ this.label }</div>
 				{ options }
-				{ this.help && <span className="help">{ this.help }</span> }
+				{ this.help && <span className='help'>{ this.help }</span> }
 			</div>
 		);
