+++ src/app/components/main/components/group/components/fields/slider/index.tsx
@@ -9,10 +9,10 @@ class SliderField extends React.Component {
 
 		return (
 			<div className={`field field-slider key-${this.field.key}`}>
-				<div className="field-label">{ this.label }</div>
-				<input type="range" onChange={ this.onChange.bind(this) } min={ this.min } max={ this.max } step={ this.step } value={ this.value } aria-label={ this.label }/>
+				<div className='field-label'>{ this.label }</div>
+				<input type='range' onChange={ this.onChange.bind(this) } min={ this.min } max={ this.max } step={ this.step } value={ this.value } aria-label={ this.label }/>
 				<label>{ this.value }</label>
-				{ this.help && <span className="help">{ this.help }</span> }
+				{ this.help && <span className='help'>{ this.help }</span> }
 			</div>
 		);
